# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: English Grammar Learning API
  version: 0.1.0
  description: Contract-first local API. Feature operations remain contract-only until their roadmap phase.
  license:
    name: Private personal use
    identifier: UNLICENSED
servers:
  - url: http://localhost:3001/api/v1
security:
  - sessionCookie: []
paths:
...
  /sessions:
    post:
      operationId: startSession
      tags: [sessions]
      summary: Start a learning session
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          description: No active curriculum or matching published exercises exist
...
  /sessions/{sessionId}/next:
    get:
      operationId: getNextSessionExercise
      tags: [sessions]
      summary: Get the next safe exercise presentation
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Exercise presentation without hidden evaluator data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExercisePresentation' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Resume an owned learning session from its bounded state snapshot
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Session state and current safe exercise presentation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionStateView' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sessions/{sessionId}/complete:
    post:
      operationId: completeSession
      tags: [sessions]
      summary: Complete a session after every item is completed
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Immutable session summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionSummaryView' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: english_session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  responses:
...
    Conflict:
      description: Conflicting state or reused idempotency key with a different payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource was not found or is not owned by the current user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    StartSessionRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode: { type: string, enum: [DAILY, FOCUSED, REVIEW] }
        grammarPointIds:
          type: array
          maxItems: 20
          items: { type: string, format: uuid }
        targetMinutes: { type: integer, minimum: 2, maximum: 30, default: 10 }
...
    ExercisePresentation:
      type: object
      additionalProperties: false
      required:
        [
          sessionItemId,
          exerciseId,
          type,
          contextVi,
          instructionVi,
          sourceTextVi,
          promptPayload,
          targets,
          attemptLimit,
        ]
      properties:
        sessionItemId: { type: string, format: uuid }
        exerciseId: { type: string, format: uuid }
        type: { type: string }
        contextVi: { type: string }
        instructionVi: { type: string }
        sourceTextVi: { type: string }
        promptPayload: { type: object }
        targets:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              [
                code,
                title,
                cefr,
                learningObjectiveVi,
                formPatterns,
                meaningUses,
                usageNotes,
                rules,
                examples,
              ]
            properties:
              code: { type: string }
              title: { type: string }
              cefr: { type: string, enum: [A1, A2, B1, B2, C1, C2] }
              learningObjectiveVi: { type: string }
              formPatterns:
                type: array
                items: { type: string }
              meaningUses:
                type: array
                items: { type: string }
              usageNotes:
                type: array
                items: { type: string }
              rules:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [code, type, description]
                  properties:
                    code: { type: string }
                    type: { type: string }
                    description: { type: string }
              examples:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [type, english, vietnamese, explanationVi]
                  properties:
                    type: { type: string }
                    english: { type: string }
                    vietnamese: { type: string }
                    explanationVi: { type: string }
        attemptLimit: { type: integer, minimum: 1 }
        grammarPointIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, format: uuid }
        targetMinutes: { type: integer, minimum: 5, maximum: 60, default: 10 }
    SessionView:
      type: object
      additionalProperties: false
      required: [id, status, mode, startedAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, COMPLETED, ABANDONED] }
        mode: { type: string, enum: [DAILY, FOCUSED, REVIEW] }
        startedAt: { type: string, format: date-time }
    SessionProgress:
      type: object
      additionalProperties: false
      required: [total, completed, remaining]
      properties:
        total: { type: integer, minimum: 0 }
        completed: { type: integer, minimum: 0 }
        remaining: { type: integer, minimum: 0 }
    SessionSummaryView:
      type: object
      additionalProperties: false
      required:
        [
          sessionId,
          totalItems,
          completedItems,
          acceptedItems,
          retryItems,
          durationSeconds,
          completedAt,
        ]
      properties:
        sessionId: { type: string, format: uuid }
        totalItems: { type: integer, minimum: 0 }
        completedItems: { type: integer, minimum: 0 }
        acceptedItems: { type: integer, minimum: 0 }
        retryItems: { type: integer, minimum: 0 }
        durationSeconds: { type: integer, minimum: 0 }
        completedAt: { type: string, format: date-time }
    SessionStateView:
      type: object
      additionalProperties: false
      required: [id, status, mode, startedAt, completedAt, progress, currentItem, summary]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [ACTIVE, COMPLETED, ABANDONED] }
        mode: { type: string, enum: [DAILY, FOCUSED, REVIEW] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        progress: { $ref: '#/components/schemas/SessionProgress' }
        currentItem:
          oneOf:
            - $ref: '#/components/schemas/ExercisePresentation'
            - type: 'null'
        summary:
          oneOf:
            - $ref: '#/components/schemas/SessionSummaryView'
            - type: 'null'
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
            message: { type: string }
            requestId: { type: string, format: uuid }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [reason]
                properties:
                  field: { type: string }
                  reason: { type: string }